agents/chat_agent/requirements.txt
# Python requirements for my_fastapi_agent

# Agent Framework packages (install from local monorepo)
# agent-framework-core
agent-framework-devui
agent-framework
agent-framework-redis
redis

# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Required by devui
openai>=1.40.0
pydantic>=2.0.0

# Config and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0

# Optional: JWT verification in production
# python-jose[cryptography]>=3.3.0
